+++ frontend/src/components/features/account/delete/Responsive.tsx
@@ -1,10 +1,10 @@
 'use client'
 
 import { useMediaQuery } from '~/hooks/useMediaQuery'
-import { DeleteAccountDialog } from './DeleteAccountDialog'
-import { DeleteAccountDrawer } from './DeleteAccountDrawer'
+import DeleteAccountDialog from './Dialog'
+import DeleteAccountDrawer from './Drawer'
 
-export function DeleteAccount() {
+export default function DeleteAccount() {
   const isDesktop = useMediaQuery('(min-width: 768px)')
   return (
     <>
